+++ js/smooth_scroll.js
@@ -1,0 +1,15 @@
+// Lenis + Gsap smooth Scrolling
+export  function smoothScroll() {
+    gsap.registerPlugin(ScrollTrigger);
+  
+    // Lenis smooth website scrolling---gsap
+    const lenis = new Lenis()
+    
+    lenis.on('scroll', ScrollTrigger.update)
+    
+    gsap.ticker.add((time) => {
+        lenis.raf(time * 1000)
+    })
+    
+    gsap.ticker.lagSmoothing(0)
+  }
